renamed ROS-related files
