+++ aegis_gym/robot/__init__.py
@@ -1,2 +1,2 @@
-from .ros_interface_factory import ROSInterfaceType, get_ros_interface  # noqa: F401
-from .base_ros_interface import BaseROSInterface  # noqa: F401
+from .robot_commander_factory import ROSInterfaceType, get_ros_interface  # noqa: F401
+from .robot_commander_interface import BaseROSInterface  # noqa: F401
